Add pre-commit hooks and pyproject.toml file
This commit adds pre-commit hooks to the repository. The hooks are
configured in the .pre-commit-config.yaml file. The pyproject.toml file
is also added to the repository. This file is used to configure the
project dependencies and the version of Python to use. At the moment,
the project dependencies are not specified in the pyproject.toml file,
but only the ruff configurations are.
Ruff replace PyLint as the linter in teh project.
Ruff is also the formatter of the project (Black style).