Added Measurements class and module
The Measurements class is used to read and store measurements from one or
multiple files. The class is used to read the measurements from the files and
store them in a pandas DataFrame. To use the class it is recommended to use a
configuration file in YAML format, that specifies the files, functions, and
filters to use to retrieve the data. The Measurements class, at the moment, is
composed of only 3 attributes:
    - meta: a dictionary that stores the metadata of the measurements
    - data: a pandas DataFrame that stores the measurements
    - multi_target: a boolean that specifies if the measurements are
        relative to multiple targets or not.

This change removes the older data_handling module, and changes the future
aspect of the application since I have decided to disentangle the measurements
from the models and/or statistics relate to them. How this will be handle is
still be completely defined.