FitModel and FitModelFactory Classes.
FitModel is a class that represents a model that can be fit to data. It can use
multiple Model-like models and combine them to obtain the scientific model
needed.
FitModelFactory is a class that can create FitModel instances. It is useful
since it can be configured from a file and can be used to create multiple
FitModel instances.